commit to backup

+++ install.sh
@@ -7,9 +7,9 @@ create_symlink() {
 	echo "... $1"
 
 	if [ $# -eq 2 ]; then
-		ln -s -f -F ~/.dotfiles/$1 ~/$2
+		ln -s -f -F ~/.dotfiles/"$1" ~/"$2"
 	else
-		ln -s -f -F ~/.dotfiles/$1 ~/$1
+		ln -s -f -F ~/.dotfiles/"$1" ~/"$1"
 	fi
 }
 
@@ -19,13 +19,13 @@ echo "Creating configuration symlinks..."
 # neovim (a recent version is needed, >0.7)
 #create_symlink .vimrc
 #rm -rf ~/.vim && create_symlink vim .vim
-mkdir -p $HOME/.config/nvim
+mkdir -p "$HOME/.config/nvim"
 #ln -s ~/.dotfiles/.vimrc ~/.config/nvim/init.vim
 #ln -s ~/.dotfiles/nvim/init.lua ~/.config/nvim/init.lua
 
 # Cheat
 # TODO add customisations to cheat
-mkdir -p $HOME/.config/cheat
+mkdir -p "$HOME/.config/cheat"
 ln -s ~/.dotfiles/cheat/conf.yml ~/.config/cheat/conf.yml
 
 # zsh
@@ -45,6 +45,10 @@ create_symlink .tmux.conf
 # ideavim
 # ln -s ~/.dotfiles/.ideavimrc ~/.ideavimrc
 create_symlink .ideavimrc
+
+# ghostty
+mkdir -p "$HOME/.config/ghostty"
+ln -s ~/.dotfiles/ghostty/config ~/.config/ghostty/config
 
 # fish (don't use anymore)
 #mv ~/.config/fish ~/.config/fish_backup 2> /dev/null || true
@@ -76,8 +80,8 @@ echo "Installing tools..."
 
 # Mac via brew
 if [[ "$OSTYPE" == "darwin"* ]]; then
+	brew install --cask ghostty
 	brew install \
-		fish \
 		neovim \
 		eza \
 		cheat \
@@ -93,8 +97,9 @@ if [[ "$OSTYPE" == "darwin"* ]]; then
 		lazygit \
 		go-task/tap/go-task \
 		tealdeer
+	#fish \
 
-	$(brew --prefix)/opt/fzf/install
+	"$(brew --prefix)"/opt/fzf/install
 
 	#zim
 	curl -fsSL https://raw.githubusercontent.com/zimfw/install/master/install.zsh | zsh
@@ -109,16 +114,15 @@ fi
 #    exit 1
 #fi
 
-## Detect system architecture
-#arch=$(uname -m)
-## Set variable based on architecture
-#if [[ "$arch" == "x86_64" || "$arch" == "i386" || "$arch" == "i686" ]]; then
-#    arch_type="x86"
-#elif [[ "$arch" == "armv7l" || "$arch" == "aarch64" ]]; then
-#    arch_type="arm"
-#else
-#    arch_type="unknown"
-#fi
+### Detect system architecture
+arch=$(uname -m)
+if [[ "$arch" == "x86_64" || "$arch" == "i386" || "$arch" == "i686" ]]; then
+	arch_type="x86"
+elif [[ "$arch" == "armv7l" || "$arch" == "aarch64" ]]; then
+	arch_type="arm"
+else
+	arch_type="unknown"
+fi
 
 # Ubuntu via apt
 # distro package version of neovim is not adequate on ubuntu
@@ -129,23 +133,38 @@ if [ -x "$(command -v apt-get)" ]; then
 	sudo apt install fzf ripgrep jq bat hexyl tmux fd-find
 	curl -s "https://get.sdkman.io" | bash
 
-	# Nvim
-	curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux-x86_64.tar.gz
-	sudo rm -rf /opt/nvim
-	sudo tar -C /opt -xzf nvim-linux-x86_64.tar.gz
+	# Download and install nvm:
+	curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.3/install.sh | bash
+	# in lieu of restarting the shell, source '.' the newly created nvm.sh
+	\. "$HOME/.nvm/nvm.sh"
+	# Download and install Node.js:
+	nvm install 22
+	# Verify the Node.js version:
+	#node -v
+	#nvm current
 
-	# lazygit
-	LAZYGIT_VERSION=$(curl -s "https://api.github.com/repos/jesseduffield/lazygit/releases/latest" | grep -Po '"tag_name": "v\K[^"]*')
-	curl -Lo lazygit.tar.gz "https://github.com/jesseduffield/lazygit/releases/latest/download/lazygit_${LAZYGIT_VERSION}_Linux_x86_64.tar.gz"
-	tar xf lazygit.tar.gz lazygit
-	sudo install lazygit /usr/local/bin
-	rm lazygit.tar.gz lazygit
+	if [[ "$arch_type" == "x86" ]]; then
+		# Nvim
+		curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux-x86_64.tar.gz
+		sudo rm -rf /opt/nvim
+		sudo tar -C /opt -xzf nvim-linux-x86_64.tar.gz
+
+		# lazygit
+		LAZYGIT_VERSION=$(curl -s "https://api.github.com/repos/jesseduffield/lazygit/releases/latest" | grep -Po '"tag_name": "v\K[^"]*')
+		curl -Lo lazygit.tar.gz "https://github.com/jesseduffield/lazygit/releases/latest/download/lazygit_${LAZYGIT_VERSION}_Linux_x86_64.tar.gz"
+		tar xf lazygit.tar.gz lazygit
+		sudo install lazygit /usr/local/bin
+		rm lazygit.tar.gz lazygit
+
+	else
+		echo "WARNING: Not x86 arch - You likey need to install Nvim and lazygit"
+	fi
 
 	sudo snap install dust
 
 	# Change shell to zsh
 	# chsh might not work (had issues on ubuntu), so change the shell to /bin/zsh in /etc/passwd
-	#chsh -s $(which zsh)
+	sudo chsh -s "$(which zsh)"
 fi
 
 # Vim-pane & tmux-split integration
@@ -158,7 +177,7 @@ echo "Setting up vim plugins..."
 vim +PlugInstall +qall
 #cd vim/plugged/YouCompleteMe
 #./install.py --all
-cd -
+#cd -
 echo "Done."
 
 echo "Setting up tmux plugins"
@@ -175,7 +194,7 @@ echo "Required actions:"
 echo "   - Install tmux plugins with 'tmux-prefix + I' (capital i, as in Install, default tmux prefix is Ctrl-b)."
 echo "   - Install vim plugins within vim with :Lazy"
 if [ -x "$(command -v apt-get)" ]; then
-	echo "Shell is: " $SHELL
+	echo "Shell is:  $SHELL"
 	echo "IF above isn't /bin/zsh, update your default shell: 'sudo /etc/passwd' and change to /bin/zsh"
 fi
 echo "Finished. Restart your shell for changes to take effect."
